Moving TestResult to file + cleanup APIWrapper
The class sis super small but standalones (well - it inherits from
HostData but still).

Also decided to move the APIWrapper code out from that sub-namepsace
back onto the Symantec.CWoC name. Saving a few lines and simplifying
calls to those classes (that an well distinguished with the exception of
the Timer class -> that's a Q on my next task :D)